fix: don't require internet on subsequent runs
