+++ lib/pre_build.ts
@@ -6,7 +6,7 @@ import { getCargoWorkspace, WasmCrate } from "./manifest.ts";
 import { verifyVersions } from "./versions.ts";
 import { BindgenOutput, generateBindgen } from "./bindgen.ts";
 import { pathExists } from "./helpers.ts";
-import { fetchWithRetries } from "../loader/fetch.js";
+import { fetchWithRetries } from "../loader/fetch.ts";
 export type { BindgenOutput } from "./bindgen.ts";
 
 export interface PreBuildOutput {
@@ -133,7 +133,7 @@ async function getBindingJsOutput(
 // deno-fmt-ignore-file`;
   const genText = bindgenOutput.js.replace(
     /\bconst\swasm_url\s.+/ms,
-    await getLoaderText(args, crate, bindgenOutput, bindingJsPath),
+    getLoaderText(args, crate, bindgenOutput, bindingJsPath),
   );
   const bodyText = await getFormattedText(`
 // source-hash: ${sourceHash}
@@ -195,7 +195,7 @@ ${genText}
   }
 }
 
-async function getLoaderText(
+function getLoaderText(
   args: CheckCommand | BuildCommand,
   crate: WasmCrate,
   bindgenOutput: BindgenOutput,
@@ -205,14 +205,14 @@ async function getLoaderText(
     case "sync":
       return getSyncLoaderText(bindgenOutput);
     case "async":
-      return await getAsyncLoaderText(
+      return getAsyncLoaderText(
         crate,
         bindgenOutput,
         false,
         bindingJsPath,
       );
     case "async-with-cache":
-      return await getAsyncLoaderText(
+      return getAsyncLoaderText(
         crate,
         bindgenOutput,
         true,
@@ -299,7 +299,7 @@ function parseRelativePath(
   return path.isAbsolute(relativeFromTo) ? specifier : relativeFromTo;
 }
 
-async function getAsyncLoaderText(
+function getAsyncLoaderText(
   crate: WasmCrate,
   bindgenOutput: BindgenOutput,
   useCache: boolean,
@@ -307,13 +307,7 @@ async function getAsyncLoaderText(
 ) {
   const exportNames = getExportNames(bindgenOutput);
 
-  const fetchContents = (await fetchModuleContents("../loader/fetch.js"))
-    .replaceAll("export async function", "async function");
-  const loaderContents = (await fetchModuleContents("../loader/mod.js"))
-    .replace(`import { fetchWithRetries } from "./fetch.js";`, "")
-    .replaceAll("export class", "class");
-
-  let loaderText = fetchContents + "\n" + loaderContents + "\n";
+  let loaderText = getWasmbuildLoaderText();
 
   let cacheText = "";
   if (useCache) {
@@ -387,22 +381,175 @@ export function isInstantiated() {
 `;
 
   return loaderText;
+
+  function getWasmbuildLoaderText() {
+return `/**
+* @callback WasmBuildDecompressCallback
+* @param {Uint8Array} compressed
+* @returns {Uint8Array} decompressed
+*/
+
+/**
+* @callback WasmBuildCacheCallback
+* @param {URL} url
+* @param {WasmBuildDecompressCallback | undefined} decompress
+* @returns {Promise<URL |Uint8Array>}
+*/
+
+/**
+* @typedef WasmBuildLoaderOptions
+* @property {WebAssembly.Imports | undefined} imports - The Wasm module's imports.
+* @property {WasmBuildCacheCallback} [cache] - A function that caches the Wasm module to
+* a local path so that a network request isn't required on every load.
+*
+* Returns an ArrayBuffer with the bytes on download success, but cache save failure.
+*/
+
+class WasmBuildLoader {
+ /** @type {WasmBuildLoaderOptions} */
+ #options;
+ /** @type {Promise<WebAssembly.WebAssemblyInstantiatedSource> | undefined} */
+ #lastLoadPromise;
+ /** @type {WebAssembly.WebAssemblyInstantiatedSource | undefined} */
+ #instantiated;
+
+ /** @param {WasmBuildLoaderOptions} options */
+ constructor(options) {
+   this.#options = options;
+ }
+
+ /** @returns {WebAssembly.Instance | undefined} */
+ get instance() {
+   return this.#instantiated?.instance;
+ }
+
+ /** @returns {WebAssembly.Module | undefined} */
+ get module() {
+   return this.#instantiated?.module;
+ }
+
+ /**
+  * @param {URL} url
+  * @param {WasmBuildDecompressCallback | undefined} decompress
+  * @returns {Promise<WebAssembly.WebAssemblyInstantiatedSource>}
+  */
+ load(
+   url,
+   decompress,
+ ) {
+   if (this.#instantiated) {
+     return Promise.resolve(this.#instantiated);
+   } else if (this.#lastLoadPromise == null) {
+     this.#lastLoadPromise = (async () => {
+       try {
+         this.#instantiated = await this.#instantiate(url, decompress);
+         return this.#instantiated;
+       } finally {
+         this.#lastLoadPromise = undefined;
+       }
+     })();
+   }
+   return this.#lastLoadPromise;
+ }
+
+ /**
+  * @param {URL} url
+  * @param {WasmBuildDecompressCallback | undefined} decompress
+  */
+ async #instantiate(url, decompress) {
+   const imports = this.#options.imports;
+   if (this.#options.cache != null && url.protocol !== "file:") {
+     try {
+       const result = await this.#options.cache(
+         url,
+         decompress ?? ((bytes) => bytes),
+       );
+       if (result instanceof URL) {
+         url = result;
+         decompress = undefined; // already decompressed
+       } else if (result != null) {
+         return WebAssembly.instantiate(result, imports);
+       }
+     } catch {
+       // ignore if caching ever fails (ex. when on deploy)
+     }
+   }
+
+   const isFile = url.protocol === "file:";
+
+   // make file urls work in Node via dnt
+   const isNode =
+     (/** @type {any} */ (globalThis)).process?.versions?.node != null;
+   if (isFile && typeof Deno !== "object") {
+     throw new Error(
+       "Loading local files are not supported in this environment",
+     );
+   }
+   if (isNode && isFile) {
+     // the deno global will be shimmed by dnt
+     const wasmCode = await Deno.readFile(url);
+     return WebAssembly.instantiate(
+       decompress ? decompress(wasmCode) : wasmCode,
+       imports,
+     );
+   }
+
+   switch (url.protocol) {
+     case "file:":
+     case "https:":
+     case "http:": {
+       const wasmResponse = await fetchWithRetries(url);
+       if (decompress) {
+         const wasmCode = new Uint8Array(await wasmResponse.arrayBuffer());
+         return WebAssembly.instantiate(decompress(wasmCode), imports);
+       }
+       if (
+         isFile ||
+         wasmResponse.headers.get("content-type")?.toLowerCase()
+           .startsWith("application/wasm")
+       ) {
+         return WebAssembly.instantiateStreaming(
+           // Cast to any so there's no type checking issues with dnt
+           // (https://github.com/denoland/wasmbuild/issues/92)
+           /** @type {any} */ (wasmResponse),
+           imports,
+         );
+       } else {
+         return WebAssembly.instantiate(
+           await wasmResponse.arrayBuffer(),
+           imports,
+         );
+       }
+     }
+     default:
+       throw new Error(\`Unsupported protocol: \${url.protocol}\`);
+   }
+ }
 }
 
-async function fetchModuleContents(path: string) {
-  const url = import.meta.resolve(path);
-  const dataResponse = await fetchWithRetries(url);
-  if (!dataResponse.ok) {
-    throw new Error(
-      `Failed fetching ${url}: ${dataResponse.statusText} - ${await dataResponse
-        .text()}`,
-    );
+/** @param {URL | string} url */
+async function fetchWithRetries(url, maxRetries = 5) {
+ let sleepMs = 250;
+ let iterationCount = 0;
+ while (true) {
+   iterationCount++;
+   try {
+     const res = await fetch(url);
+     if (res.ok || iterationCount > maxRetries) {
+       return res;
+     }
+   } catch (err) {
+     if (iterationCount > maxRetries) {
+       throw err;
+     }
+   }
+   console.warn(\`Failed fetching. Retrying in \${sleepMs}ms...\`);
+   await new Promise((resolve) => setTimeout(resolve, sleepMs));
+   sleepMs = Math.min(sleepMs * 2, 10_000);
+ }
+}
+`;
   }
-  return (await dataResponse.text())
-    .replace(
-      "// Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.\n",
-      "",
-    );
 }
 
 function getExportNames(bindgenOutput: BindgenOutput) {
